feat(form): using actions to handle form submission
