+++ src/routes/+page.server.ts
@@ -1,4 +1,6 @@
 import { getTeams } from '$lib/linear';
+import { fail, type Actions } from '@sveltejs/kit';
+import { validateForm } from '../utils/validateForm';
 
 export const load = async () => {
   try {
@@ -6,5 +8,25 @@ export const load = async () => {
     return { teams };
   } catch (err) {
     console.log(err);
+  }
+};
+
+export const actions: Actions = {
+  submitReport: async ({ request }) => {
+    const formData = await request.formData();
+    const isValid = validateForm(formData);
+
+    if (!isValid) {
+      return fail(400, { message: 'Error' });
+    }
+
+    const data: any = {};
+    for (const field of formData) {
+      const [key, value] = field;
+      data[key] = value;
+    }
+
+    console.log(data);
+    console.log(isValid);
   }
 };
